[BACKLOG-38435] Code changes for moving Mail Validator from core to plugin.
