+++ engine/src/main/resources/org/pentaho/di/trans/step/messages/messages_en_US.properties
@@ -147,7 +147,6 @@ BaseStep.TypeTooltipDesc.MondrianInput=Execute and retrieve data using an MDX qu
 BaseStep.TypeTooltipDesc.OlapInput=Execute and retrieve data using an MDX query against any XML/A OLAP datasource using olap4j
 BaseStep.Log.ErrorInThreadSleeping=Interupted while trying to put a new row in a buffer\:
 BaseStep.TypeTooltipDesc.Update=Update data in a database table based upon keys
-BaseStep.TypeTooltipDesc.MailValidator=Check if an email address is valid.
 BaseStep.TypeLongDesc.DatabaseJoin=Database join
 BaseStep.Log.AllowedRelationships=Only 1-1, 1-n, n-1 and n-n relationships are allowed\!
 BaseStep.TypeLongDesc.StepMetastructure=Metadata structure of stream
@@ -278,7 +277,6 @@ BaseStep.TypeTooltipDesc.CopyRowsToResult=Use this step to write rows to the exe
 BaseStep.TypeTooltipDesc.MapValues=Maps values of a certain field from one value to another
 BaseStepDialog.Cancel=&Cancel
 BaseStep.TypeLongDesc.SynchronizeAfterMerge=Synchronize after merge
-BaseStep.TypeLongDesc.MailValidator=Mail validator
 BaseStep.TypeTooltipDesc.Uniquerows=Remove double rows and leave only unique occurrences.\nThis works only on a sorted input.\nIf the input is not sorted, only double consecutive rows are handled correctly.
 BaseStep.TypeLongDesc.XMLOutput=XML output
 BaseStep.TypeTooltipDesc.Janino=Calculate the result of a Java Expression using Janino
