--- plugins/mail-validator-job/assemblies/plugin/pom.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.pentaho.di.plugins</groupId>
-    <artifactId>mail-validator-job-assemblies</artifactId>
-    <version>10.1.0.0-SNAPSHOT</version>
-  </parent>
-
-
-  <artifactId>mail-validator-job-plugins</artifactId>
-  <version>10.1.0.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-
-  <name>PDI Mail Validator Job Plugins Distribution</name>
-  <description></description>
-  <properties>
-    <resources.directory>${project.basedir}/src/main/resources</resources.directory>
-  </properties>
-  <dependencies>
-    <dependency>
-      <groupId>org.pentaho.di.plugins</groupId>
-      <artifactId>mail-validator-job-plugins-impl</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.pentaho.di.plugins</groupId>
-      <artifactId>mail-validator-job-plugins-ui</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-dependency-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>add-libs</id>
-            <phase>generate-resources</phase>
-            <goals>
-              <goal>copy-dependencies</goal>
-            </goals>
-            <configuration>
-              <outputDirectory>${project.build.directory}/assembly/lib</outputDirectory>
-              <includeScope>compile</includeScope>
-              <excludeArtifactIds>mail-validator-job-plugins-impl,mail-validator-job-plugins-ui</excludeArtifactIds>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
