--- plugins/mail-validator-job/assemblies/plugin/src/assembly/assembly.xml
@@ -1,41 +1,0 @@
-<assembly xmlns="http://maven.apache.org/ASSEMBLY/2.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-          xsi:schemaLocation="http://maven.apache.org/ASSEMBLY/2.0.0 http://maven.apache.org/xsd/assembly-2.0.0.xsd">
-  <id>mail-validator-job-plugin</id>
-  <formats>
-    <format>zip</format>
-  </formats>
-  <baseDirectory>mail-validator-job-plugins</baseDirectory>
-
-  <files>
-    <file>
-      <source>${project.basedir}/src/main/resources/version.xml</source>
-      <outputDirectory>.</outputDirectory>
-      <filtered>true</filtered>
-    </file>
-  </files>
-
-  <fileSets>
-    <!-- the staging dir -->
-    <fileSet>
-      <directory>${project.build.directory}/assembly</directory>
-      <outputDirectory>.</outputDirectory>
-    </fileSet>
-
-    <fileSet>
-      <directory>${resources.directory}</directory>
-      <outputDirectory>.</outputDirectory>
-      <filtered>true</filtered>
-    </fileSet>
-  </fileSets>
-
-  <dependencySets>
-    <dependencySet>
-      <outputDirectory>.</outputDirectory>
-      <includes>
-        <include>org.pentaho.di.plugins:mail-validator-job-plugins-impl:jar</include>
-        <include>org.pentaho.di.plugins:mail-validator-job-plugins-ui:jar</include>
-      </includes>
-      <useProjectArtifact>false</useProjectArtifact>
-    </dependencySet>
-  </dependencySets>
-</assembly>
