--- plugins/mail-validator-job/assemblies/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.pentaho.di.plugins</groupId>
-    <artifactId>mail-validator-job</artifactId>
-    <version>10.1.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>mail-validator-job-assemblies</artifactId>
-  <version>10.1.0.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-
-  <name>PDI Mail Validator Job Plugins Assemblies</name>
-
-  <modules>
-    <module>plugin</module>
-  </modules>
-</project>
