--- plugins/mail-validator-job/pom.xml
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.pentaho.di.plugins</groupId>
-    <artifactId>pdi-plugins</artifactId>
-    <version>10.1.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>mail-validator-job</artifactId>
-  <version>10.1.0.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-
-  <name>PDI Mail Validator Job Plugins</name>
-  <description>This plugin is for pentaho Data integration (ETL) a.k.a kettle</description>
-  <url>http://www.pentaho.com/</url>
-
-  <modules>
-    <module>impl</module>
-    <module>ui</module>
-    <module>assemblies</module>
-  </modules>
-
-  <licenses>
-    <license>
-      <name>Apache License, Version 2.0</name>
-      <url>https://www.apache.org/licenses/LICENSE-2.0.txt</url>
-      <distribution>repo</distribution>
-      <comments>A business-friendly OSS license</comments>
-    </license>
-  </licenses>
-
-  <properties>
-    <org.eclipse.swt.version>4.6</org.eclipse.swt.version>
-    <jface.version>3.3.0-I20070606-0010</jface.version>
-    <pdi.version>10.1.0.0-SNAPSHOT</pdi.version>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>pentaho-kettle</groupId>
-        <artifactId>kettle-engine</artifactId>
-        <version>${pdi.version}</version>
-        <scope>provided</scope>
-      </dependency>
-      <dependency>
-        <groupId>pentaho-kettle</groupId>
-        <artifactId>kettle-core</artifactId>
-        <version>${pdi.version}</version>
-        <scope>provided</scope>
-      </dependency>
-      <dependency>
-        <groupId>pentaho-kettle</groupId>
-        <artifactId>kettle-ui-swt</artifactId>
-        <version>${pdi.version}</version>
-        <scope>provided</scope>
-      </dependency>
-
-      <!-- SWT is required to compile,inorder to work with architecture -->
-      <dependency>
-        <groupId>org.eclipse.swt</groupId>
-        <artifactId>org.eclipse.swt.gtk.linux.x86_64</artifactId>
-        <version>${org.eclipse.swt.version}</version>
-        <scope>provided</scope>
-        <exclusions>
-          <exclusion>
-            <artifactId>*</artifactId>
-            <groupId>*</groupId>
-          </exclusion>
-        </exclusions>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse</groupId>
-        <artifactId>jface</artifactId>
-        <version>${jface.version}</version>
-        <scope>provided</scope>
-        <exclusions>
-          <exclusion>
-            <artifactId>*</artifactId>
-            <groupId>*</groupId>
-          </exclusion>
-        </exclusions>
-      </dependency>
-      <!--test dependencies-->
-      <dependency>
-       <groupId>pentaho-kettle</groupId>
-       <artifactId>kettle-core</artifactId>
-       <version>${pdi.version}</version>
-       <classifier>tests</classifier>
-       <scope>test</scope>
-      </dependency>
-      <dependency>
-         <groupId>pentaho-kettle</groupId>
-         <artifactId>kettle-engine</artifactId>
-         <version>${pdi.version}</version>
-         <classifier>tests</classifier>
-         <scope>test</scope>
-       </dependency>
-
-      <dependency>
-        <groupId>com.healthmarketscience.jackcess</groupId>
-        <artifactId>jackcess</artifactId>
-        <version>1.2.6</version>
-        <scope>compile</scope>
-        <exclusions>
-          <exclusion>
-            <artifactId>*</artifactId>
-            <groupId>*</groupId>
-          </exclusion>
-        </exclusions>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-</project>
