+++ plugins/mail-validator/assemblies/plugin/src/main/resources/version.xml
