+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/job/entries/mailvalidator/messages/messages_fr_FR.properties
