+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/job/entries/mailvalidator/messages/messages_ko_KR.properties
