+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/trans/steps/mailvalidator/messages/messages_en_US.properties
@@ -1,0 +1,95 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.MailValidator' in locale 'en_US'
+#
+#
+#Thu Jun 19 19:10:46 CEST 2008
+
+BaseStep.TypeLongDesc.MailValidator=Mail validator
+BaseStep.TypeTooltipDesc.MailValidator=Check if an email address is valid.
+MailValidatorDialog.Shell.Title=Mail validator
+MailValidatorDialog.FailedToGetFields.DialogMessage=Unable to get fields from previous steps because of an error
+MailValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Unexpected error reading step information from the repository
+MailValidatorDialog.Stepname.Label=Step name 
+MailValidator.Log.UnexpectedError=Unexpected error in ''
+MailValidatorDialog.emailFieldName.Label=email fieldname  
+MailValidatorDialog.FailedToGetFields.DialogTitle=Get fields failed
+MailValidator.LineNumber=linenr {0}
+MailValidatorMeta.CheckResult.TableFieldOK=Table fieldname is provided
+MailValidator.ErrorInStepRunning=Because of an error, this step can''t continue\: 
+MailValidator.Log.ErrorFindingField=Error finding field\: 
+MailValidatorMeta.Exception.UnableToSaveStepInfo=Unable to save step information to the repository for id_step\=
+MailValidatorMeta.CheckResult.CouldNotReadFields=Couldn''t read fields from the previous step.
+
+MailValidatorMeta.CheckResult.NoInpuReceived=No input received from other steps\!
+MailValidator.Exception.CouldnotFindField=Couldn''t find field ''{0}'' in row\!
+MailValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Step is receiving info from other steps.
+MailValidatorDialog.ResultField.Label=Result fieldname
+MailValidatorDialog.ResultField.Tooltip=Result fieldname
+
+MailValidator.Error.FilenameFieldMissing=Filename is missing\!
+MailValidatorMeta.CheckResult.ResultFieldMissing=Result fieldname is missing\!
+MailValidatorMeta.CheckResult.ResultFieldOk=Result fieldname was specified.
+
+MailValidatorMeta.CheckResult.ErrorOccurred=An error occurred\: 
+
+MailValidatorMeta.Exception.UnableToReadStepInfo=Unable to read step information from XML
+MailValidatorMeta.CheckResult.eMailFieldMissing=Email field name is missing\!
+MailValidatorMeta.CheckResult.eMailFieldOK=Email field name was specified
+MailValidatorDialog.ResultGroup.label=Result
+MailValidatorDialog.ResultAsString.Label=Result is a string
+MailValidatorDialog.ResultAsString.Tooltip=Result is a string
+MailValidatorDialog.ResultStringTrueField.Label=Email is valid
+MailValidatorDialog.ResultStringTrueField.Tooltip=display when email address is valid
+MailValidatorDialog.ResultStringFalseField.Label=Email is not valid
+MailValidatorDialog.ResultStringFalseField.Tooltip=display when email is not valid
+MailValidatorMeta.CheckResult.eMailValidMsgMissing=eMail valid message is missing!
+MailValidatorMeta.CheckResult.eMailValidMsgOk=eMail valid message was specified.
+MailValidatorMeta.CheckResult.eMailNotValidMsgMissing=eMail not valid message is missing!
+MailValidatorMeta.CheckResult.eMailNotValidMsgOk=eMail not valid message was specified.
+MailValidatorDialog.ErrorMsgField.Label=Errors field
+MailValidatorDialog.ErrorMsgField.Tooltip=Errors field
+MailValidator.Error.ResultFieldNameMissing=Result fieldname is missing or empty!
+MailValidatorDialog.SettingsGroup.Label=Settings
+MailValidatorDialog.SMTPCheck.Label=SMTP check?
+MailValidatorDialog.SMTPCheck.Tooltip=If you select this option, PDI will perform a SMTP check.\nBy default, only the mail structure will be checked.
+MailValidatorDialog.eMailSenderField.Label=Email sender
+MailValidatorDialog.eMailSenderField.Tooltip=This address will be used when performing SMTP check.
+MailValidatorDialog.TimeOutField.Label=Time out (milliseconds)
+MailValidatorDialog.TimeOutField.Tooltip=In order to perform a SMTP check, PDI will open a socket.\nSpecify here the timeout in milliseconds.
+MailValidatorDialog.DefaultSMTPField.Label=Default SMTP server
+MailValidatorDialog.DefaultSMTPField.Tooltip=When PDI perform a SMTP check,\nit extracts exhangers list from email's domain name.\nDepending on the size of the list, the process can last.\nIf you know which SMTP server to query, please specify it as default SMTP server\nin that case, PDI will query only this server.
+MailValidatorMeta.CheckResult.eMailSenderMissing=Email sender is missing!
+MailValidatorMeta.CheckResult.eMailSenderOk=Email sender was specified.
+MailValidatorDialog.dynamicDefaultSMTP.Label=dynamic default SMTP?
+MailValidatorDialog.dynamicDefaultSMTP.Tooltip=Select this option if you want to get dynamically\ndefault SMTP server.You must select below the field that contain value.
+MailValidatorDialog.defaultSMTPField.Label=Default SMTP field
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldMissing=Dynamic default SMTP field is missing!
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldOk=Dynamic default SMTP field was specified.
+MailValidator.Error.EmailValidMsgMissing=Email valid message is empty!
+MailValidatorMeta.CheckResult.EmailValidMsgOk=Email valid message was specified.
+MailValidator.Error.EmailNotValidMsgMissing=Email not valid message is empty!
+MailValidatorMeta.CheckResult.EmailNotValidMsgOk=Email not valid message was specified.
+MailValidator.Error.EMailValidMsgMissing=Email valid message is empty!
+MailValidator.Error.EMailNotValidMsgMissing=Email not valid message is empty!
+MailValidator.NoMatchName=No match for name '{0}' 
+MailValidator.MalformedAddress=Malformed address {0}!
+MailValidator.ClassName=Email Validator
+MailValidator.NoMachinesInDomain=We do not find machines for domain : {0}!
+MailValidator.ErrorGettingMachinesInDomain=Error getting machines : {0}
+MailValidator.ExchangersFound=We found {0} exchanger(s).
+MailValidator.TryingExchanger=Trying Exchanger : {0} ...
+MailValidator.InvalidHeader=Invalid header
+MailValidator.SayHello= 1°) Let''s say Hello to server : EHLO {0}
+MailValidator.CheckSender= 2°) Let''s check sender. MAIL FROM: <{0}> 
+MailValidator.ServerReplied=Server successfully replied : {0}
+MailValidator.SenderRejected=Sender rejected
+MailValidator.SenderAccepted=Sender accepted : {0}
+MailValidator.CheckReceiver=3°) Let''s check receiver. RCPT TO: <{0}>
+MailValidator.AddressNotValid=Address  : {0} is not valid!
+MailValidator.ReceiverAccepted=Receiver {0} accepted : {1}
+
+MailValidator.Error.DefaultSMTPFieldMissing=Default SMTP server is missing!
+MailValidator.Error.ResultFieldMissing=Result fieldname is missing!
+MailValidatorMeta.CheckResult.EmailNotValidMsgMissing=EMail not valid message is missing!
+MailValidatorMeta.CheckResult.EmailValidMsgMissing=EMail valid message is missing!
+MailValidator.Error.MailEmpty=Email address is empty!
+MailValidator.ConnectingTo=Connecting to {0} on port {1} ...{2}
