+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/trans/steps/mailvalidator/messages/messages_es_AR.properties
@@ -1,0 +1,80 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.mailvalidator' in locale 'es_AR'
+#
+#
+#Fri Jun 04 15:47:47 ART 2010
+MailValidatorDialog.ResultStringTrueField.Tooltip=mostrar cuando la direcci\u00F3n es v\u00E1lida
+MailValidator.SayHello=1\u00B0) Saludo al servidor\: EHLO {0}
+MailValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=El paso est\u00E1 recibiendo informaci\u00F3n de otros pasos.
+MailValidatorDialog.ErrorMsgField.Label=Campo de errores
+MailValidator.NoMachinesInDomain=No se encuentran m\u00E1quinas para el dominio\: {0}.
+MailValidatorDialog.eMailSenderField.Tooltip=Esta direcci\u00F3n se utilizar\u00E1 al realizar la verificaci\u00F3n SMTP.
+MailValidatorDialog.ResultStringTrueField.Label=Direcci\u00F3n de correo v\u00E1lida
+MailValidatorDialog.dynamicDefaultSMTP.Tooltip=Seleccione esta opci\u00F3n si desea obtener el SMTP\n por defecto din\u00E1micamente. Debe seleccionar el campo que contendr\u00E1 el valor\n en la opci\u00F3n siguiente.
+MailValidatorMeta.CheckResult.ResultFieldOk=Se ha indicado el campo para el resultado.
+MailValidator.ServerReplied=El servidor contest\u00F3 correctamente\: {0}
+MailValidator.ExchangersFound=Se encontraron {0} intercambiadores.
+MailValidatorDialog.eMailSenderField.Label=Remitente
+MailValidatorMeta.Exception.UnableToReadStepInfo=Imposible leer informaci\u00F3n del paso desde XML.
+MailValidatorMeta.CheckResult.eMailSenderOk=Se ha indicado el remitente.
+MailValidator.Error.EMailNotValidMsgMissing=Falta el mensaje para correo no v\u00E1lido.
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldMissing=Falta el campo para el SMTP por defecto din\u00E1mico.
+MailValidatorDialog.ResultAsString.Label=El resultado es una cadena
+MailValidator.SenderRejected=Remitente rechazado
+MailValidatorMeta.CheckResult.EmailValidMsgOk=Se ha indicado el mensaje para direcciones v\u00E1lidas.
+MailValidatorMeta.CheckResult.eMailValidMsgOk=Se ha indicado el mensaje para direcciones v\u00E1lidas.
+MailValidatorDialog.defaultSMTPField.Label=Campos para SMTP por defecto
+MailValidatorDialog.dynamicDefaultSMTP.Label=\u00BFSMTP por defecto din\u00E1mico?
+MailValidator.LineNumber=NroLinea {0}
+MailValidator.Error.ResultFieldNameMissing=Falta el campo para el resultado.
+MailValidatorMeta.CheckResult.eMailFieldOK=Se ha indicado el campo para la direcci\u00F3n de correo.
+MailValidatorDialog.ResultField.Tooltip=Campo para resultado
+MailValidatorMeta.CheckResult.eMailNotValidMsgMissing=Falta el mensaje para direcciones inv\u00E1lidas.
+MailValidatorDialog.TimeOutField.Label=L\u00EDmite de tiempo (ms)
+MailValidator.TryingExchanger=Probando intercambiador\: {0}...
+MailValidator.MalformedAddress=Direcci\u00F3n malformada {0}
+MailValidatorMeta.CheckResult.eMailSenderMissing=Falta el remitente.
+MailValidatorDialog.TimeOutField.Tooltip=Para realizar la verificaci\u00F3n SMTP, PDI abrir\u00E1 un socket.\nIndique aqu\u00ED el l\u00EDmite de tiempo en milisegundos.
+MailValidatorMeta.Exception.UnableToSaveStepInfo=Imposible guardar informaci\u00F3n del paso en el repositorio para el paso con id\=
+MailValidatorDialog.ResultStringFalseField.Tooltip=mostrar cuando la direcci\u00F3n es inv\u00E1lida
+MailValidatorMeta.CheckResult.ResultFieldMissing=Falta el campo para el resultado.
+MailValidator.Error.DefaultSMTPFieldMissing=Falta el servidor SMTP por defecto.
+MailValidatorDialog.SMTPCheck.Label=\u00BFVerificaci\u00F3n SMTP?
+MailValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Error inesperado al leer informaci\u00F3n del paso desde el repositorio
+MailValidatorMeta.CheckResult.eMailNotValidMsgOk=Se ha indicado el mensaje para direcciones inv\u00E1lidas.
+MailValidatorDialog.ResultAsString.Tooltip=El resultado es una cadena
+MailValidatorMeta.CheckResult.EmailNotValidMsgMissing=Falta el mensaje para direcciones inv\u00E1lidas.
+MailValidator.ErrorGettingMachinesInDomain=Error al obtener m\u00E1quinas\: {0}
+MailValidatorMeta.CheckResult.NoInpuReceived=No se recibe entrada de otros pasos.
+MailValidatorDialog.emailFieldName.Label=Campo para direcci\u00F3n
+MailValidatorDialog.ResultStringFalseField.Label=Direcci\u00F3n de correo inv\u00E1lida
+MailValidatorMeta.CheckResult.EmailNotValidMsgOk=Se ha indicado el mensaje para direcciones inv\u00E1lidas.
+MailValidator.CheckSender=3\u00B0) Verificando el remitente. MAIL FROM\: <{0}>
+MailValidator.ConnectingTo=Conectando a {0} en el puerto {1}... {2}
+MailValidatorDialog.ResultGroup.label=Resultado
+MailValidator.Error.MailEmpty=La direcci\u00F3n de correo est\u00E1 vac\u00EDa.
+MailValidator.ErrorInStepRunning=Este paso no puede continuar debido a un error\:
+MailValidatorDialog.Stepname.Label=Nombre del paso
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldOk=Se ha indicado el campo para el SMTP por defecto din\u00E1mico.
+MailValidator.CheckReceiver=3\u00B0) Verificando el destinatario. RCPT TO\: <{0}>
+MailValidatorDialog.SettingsGroup.Label=Ajustes
+MailValidatorDialog.ResultField.Label=Campo para resultado
+MailValidatorDialog.FailedToGetFields.DialogMessage=Imposible obtener campos de pasos previos debido a un error
+MailValidator.ReceiverAccepted=El destinatario {0} acept\u00F3\: {1}
+MailValidatorDialog.FailedToGetFields.DialogTitle=Error al obtener archivos
+MailValidatorDialog.ErrorMsgField.Tooltip=Campo de errores
+MailValidatorDialog.SMTPCheck.Tooltip=Si selecciona esta opci\u00F3n, PDI realizar\u00E1 una verificaci\u00F3n SMTP.\nPor defecto, s\u00F3lo la estructura del correo ser\u00E1 verificada.
+MailValidatorDialog.DefaultSMTPField.Tooltip=Cuando PDI realiza una verificaci\u00F3n de SMTP,\nextrae la lista de intercambiadores del nombre de dominio de la direcci\u00F3n de correo.\nDependiendo del tama\u00F1o de la lista, el proceso puede demorarse.\nSi conoce qu\u00E9 servidor SMTP deber\u00EDa ser consultado, por favor ind\u00EDquelo como servidor SMTP por defecto.\nEn ese caso, PDI s\u00F3lo consultar\u00E1 a ese servidor.
+MailValidatorMeta.CheckResult.eMailValidMsgMissing=Falta el mensaje para direcciones v\u00E1lidas.
+MailValidator.SenderAccepted=Remitente aceptado\: {0}
+MailValidator.Error.ResultFieldMissing=Falta el campo para el resultado.
+MailValidator.NoMatchName=No hay coincidencias para el nombre '{0}'
+MailValidator.InvalidHeader=Encabezado inv\u00E1lido
+MailValidatorMeta.CheckResult.EmailValidMsgMissing=Falta el mensaje para direcciones v\u00E1lidas.
+MailValidatorMeta.CheckResult.eMailFieldMissing=Falta el campo para la direcci\u00F3n de correo.
+MailValidator.AddressNotValid=La direcci\u00F3n {0} es inv\u00E1lida.
+MailValidator.ClassName=Validador de Correo
+MailValidator.Error.EMailValidMsgMissing=Falta el mensaje para correo v\u00E1lido.
+MailValidatorDialog.Shell.Title=Validador de correo
+MailValidator.Error.FilenameFieldMissing=Falta el nombre de archivo.
+MailValidator.Exception.CouldnotFindField=El campo "{0}" no pudo encontrarse en la fila.
+MailValidatorDialog.DefaultSMTPField.Label=Servidor SMTP por defecto
