+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/trans/steps/mailvalidator/messages/messages_it_IT.properties
@@ -1,0 +1,80 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.mailvalidator' in locale 'it_IT'
+
+
+#Wed Sep 03 09:39:58 GMT+01:00 2008
+MailValidatorDialog.ResultStringTrueField.Tooltip=mostra quando l''indirizzo email \u00E8 valido
+MailValidator.SayHello=1\u00B0) Viene detto HELLO al server SMTP\: EHLO {0}
+MailValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Il passo sta ricevendo informazioni dagli altri passi.
+MailValidatorDialog.ErrorMsgField.Label=Campo errori
+MailValidator.NoMachinesInDomain=Non sono state trovate macchine per il dominio\: {0}\!
+MailValidatorDialog.eMailSenderField.Tooltip=Quest''indirizzo verr\u00E0 usato per il controllo dell''SMTP.
+MailValidatorDialog.ResultStringTrueField.Label=L''email \u00E8 valida
+MailValidatorDialog.dynamicDefaultSMTP.Tooltip=Selezionare quest''opzione per ottenere dinamicamente il server SMTP di default. \nOccorre selezionare qua sotto il campo che contiene il valore.
+MailValidatorMeta.CheckResult.ResultFieldOk=Il nome file del risultato \u00E8 stato specificato.
+MailValidator.ServerReplied=Il server ha risposto con successo\: {0}
+MailValidator.ExchangersFound=Trovati {0} exchanger.
+MailValidatorDialog.eMailSenderField.Label=Mittente email
+MailValidatorMeta.Exception.UnableToReadStepInfo=Impossibile leggere le informazioni del passo da XML
+MailValidatorMeta.CheckResult.eMailSenderOk=Il mittente dell''email \u00E8 stato specificato.
+MailValidator.Error.EMailNotValidMsgMissing=Manca il messaggio di email non valida\!
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldMissing=Manca il campo dinamico dell''SMTP di default\!
+MailValidatorDialog.ResultAsString.Label=Il risultato \u00E8 una stringa
+MailValidator.SenderRejected=Mittente rifiutato
+MailValidatorMeta.CheckResult.EmailValidMsgOk=Il messaggio valido della mail \u00E8 stato specificato.
+MailValidatorMeta.CheckResult.eMailValidMsgOk=Il messaggio di email valida \u00E8 stato specificato.
+MailValidatorDialog.defaultSMTPField.Label=Campo SMTP di default
+MailValidatorDialog.dynamicDefaultSMTP.Label=default SMTP dinamico?
+MailValidator.LineNumber=linea n. {0}
+MailValidator.Error.ResultFieldNameMissing=Manco od \u00E8 vuoto il nome campo del risultato\!
+MailValidatorMeta.CheckResult.eMailFieldOK=Il nome campo email \u00E8 stato specificato.
+MailValidatorDialog.ResultField.Tooltip=Nome campo del risultato
+MailValidatorMeta.CheckResult.eMailNotValidMsgMissing=Manca il messaggio di email non valida\!
+MailValidatorDialog.TimeOutField.Label=Time out (millisecondi)
+MailValidator.TryingExchanger=Viene provato l''exchanger\: {0} ...
+MailValidator.MalformedAddress=Indirizzo malformato {0}\!
+MailValidatorMeta.CheckResult.eMailSenderMissing=Manca il mittente dell''email\!
+MailValidatorDialog.TimeOutField.Tooltip=Per eseguire un controllo SMTP, PDI aprir\u00E0 un socket.\nSpecificare qua il timeout in millisecondi.
+MailValidatorMeta.Exception.UnableToSaveStepInfo=Impossibile salvare le informazioni del passo nel repository per id_step\={0}
+MailValidatorDialog.ResultStringFalseField.Tooltip=mostra quando l''indirizzo email non \u00E8 valido
+MailValidatorMeta.CheckResult.ResultFieldMissing=Manca il nome file del risultato\!
+MailValidator.Error.DefaultSMTPFieldMissing=Manca il server SMTP di default\!
+MailValidatorDialog.SMTPCheck.Label=Controllare SMTP?
+MailValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Errore inatteso durante la lettura delle informazioni del passo dal repository
+MailValidatorMeta.CheckResult.eMailNotValidMsgOk=Il messaggio di email non valida \u00E8 stato specificato.
+MailValidatorDialog.ResultAsString.Tooltip=Il risultato \u00E8 una stringa
+MailValidatorMeta.CheckResult.EmailNotValidMsgMissing=Manca il messaggio di email non valida\!
+MailValidator.ErrorGettingMachinesInDomain=Errore nella ricezione della macchine\: {0}
+MailValidatorMeta.CheckResult.NoInpuReceived=Nessun input ricevuto dagli altri passi\!
+MailValidatorDialog.emailFieldName.Label=Nome campo email
+MailValidatorDialog.ResultStringFalseField.Label=L''email non \u00E8 valida
+MailValidatorMeta.CheckResult.EmailNotValidMsgOk=Il messaggio di email non valida \u00E8 stato specificato.
+MailValidator.CheckSender=2\u00B0) Controllo il mittente. MAIL FROM\: <{0}> 
+MailValidator.ConnectingTo=Connessione a {0} sulla porta {1} ...{2}
+MailValidatorDialog.ResultGroup.label=Risultato
+MailValidator.Error.MailEmpty=L''indirizzo email \u00E8 vuoto\!
+MailValidator.ErrorInStepRunning=A causa di un errore, questo passo non pu\u00F2 continuare\: 
+MailValidatorDialog.Stepname.Label=Nome passo
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldOk=Il campo dinamico dell''SMTP di default \u00E8 stato specificato.
+MailValidator.CheckReceiver=3\u00B0) Controllo il destinatario. RCPT TO\: <{0}>
+MailValidatorDialog.SettingsGroup.Label=Impostazioni
+MailValidatorDialog.ResultField.Label=Nome campo del risultato
+MailValidatorDialog.FailedToGetFields.DialogMessage=Impossibile prelevare i campi dai precedenti passi a causa di un errore
+MailValidator.ReceiverAccepted=Ricevitore {0} accettato\: {1}
+MailValidatorDialog.FailedToGetFields.DialogTitle=Prelievo campi fallito
+MailValidatorDialog.ErrorMsgField.Tooltip=Campo errori
+MailValidatorDialog.SMTPCheck.Tooltip=Se viene selezionata quest''opzione, PDI eseguir\u00E0 un controllo SMTP.\nDi default, soltanto la struttura dell''email verr\u00E0 controllata.
+MailValidatorDialog.DefaultSMTPField.Tooltip=Quando PDI esegue un controllo SMTP, estrae la lista degli exchanger dal nome del dominio dell''email.\nA seconda della dimensione della lista, il processo pu\u00F2 durare a lungo.\nSe si conosce quale server interrogare, per favore specificarlo come server SMTP di default.\nIn questo caso, PDI interrogher\u00E0 soltanto questo server.
+MailValidatorMeta.CheckResult.eMailValidMsgMissing=Manca il messaggio di email valida\!
+MailValidator.SenderAccepted=Il mittente \u00E8 stato accettato\: {0}
+MailValidator.Error.ResultFieldMissing=Manca il nome file del risultato\!
+MailValidator.NoMatchName=Nessuna corrispondenza per il nome '{0}' 
+MailValidator.InvalidHeader=Intestazione invalida
+MailValidatorMeta.CheckResult.EmailValidMsgMissing=Manca il messaggio valido della mail\!
+MailValidatorMeta.CheckResult.eMailFieldMissing=Manca il nome campo email\!
+MailValidator.AddressNotValid=Indirizzo\: {0} non \u00E8 valido\!
+MailValidator.ClassName=Validatore email
+MailValidator.Error.EMailValidMsgMissing=Il messaggio di email valida \u00E8 vuoto\!
+MailValidatorDialog.Shell.Title=Validatore email
+MailValidator.Error.FilenameFieldMissing=Manca il nome file\!
+MailValidator.Exception.CouldnotFindField=Impossibile trovare il campo ''{0}'' nella riga\!
+MailValidatorDialog.DefaultSMTPField.Label=Server SMTP di default
