+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/trans/steps/mailvalidator/messages/messages_ko_KR.properties
@@ -1,0 +1,27 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.mailvalidator' in locale 'ko_KR'
+#
+#
+#Thu Apr 30 14:58:38 KST 2009
+MailValidatorDialog.TimeOutField.Label=\uD0C0\uC784\uC544\uC6C3
+MailValidatorDialog.ResultField.Label=\uACB0\uACFC \uD544\uB4DC\uC774\uB984
+MailValidatorDialog.ResultField.Tooltip=\uACB0\uACFC \uD544\uB4DC\uC774\uB984
+MailValidator.ClassName=Email Validator
+MailValidatorDialog.ResultStringTrueField.Label=\uC774\uBA54\uC77C\uC774 \uC720\uD6A8\uD569\uB2C8\uB2E4
+MailValidatorDialog.ResultGroup.label=\uACB0\uACFC
+MailValidatorDialog.eMailSenderField.Label=\uC774\uBA54\uC77C \uC1A1\uC2E0\uC790
+MailValidatorDialog.SMTPCheck.Label=SMTP \uCCB4\uD06C?
+MailValidatorDialog.Shell.Title=Mail validator
+MailValidatorDialog.ResultAsString.Label=\uACB0\uACFC\uB294 \uBB38\uC790\uC5F4
+MailValidatorDialog.FailedToGetFields.DialogTitle=\uD544\uB4DC \uAC00\uC838\uC624\uAE30 \uC2E4\uD328
+MailValidatorDialog.Stepname.Label=Step \uC774\uB984
+MailValidatorDialog.emailFieldName.Label=\uC774\uBA54\uC77C \uD544\uB4DC\uC774\uB984
+MailValidatorDialog.DefaultSMTPField.Label=\uAE30\uBCF8 SMTP \uC11C\uBC84
+MailValidatorDialog.SettingsGroup.Label=\uC124\uC815
+MailValidatorDialog.defaultSMTPField.Label=\uAE30\uBCF8 SMTP \uD544\uB4DC
+MailValidatorDialog.ResultAsString.Tooltip=\uACB0\uACFC\uB294 \uBB38\uC790\uC5F4\uC785\uB2C8\uB2E4
+MailValidatorDialog.ResultStringTrueField.Tooltip=\uC774\uBA54\uC77C\uC774 \uC720\uD6A8\uD558\uBA74 \uD45C\uC2DC
+MailValidatorDialog.ErrorMsgField.Label=\uC624\uB958 \uD544\uB4DC
+MailValidatorDialog.ErrorMsgField.Tooltip=\uC624\uB958 \uD544\uB4DC
+MailValidatorDialog.dynamicDefaultSMTP.Label=\uB3D9\uC801\uC778 \uAE30\uBCF8 SMTP?
+MailValidatorDialog.ResultStringFalseField.Label=\uC774\uBA54\uC77C\uC774 \uC720\uD6A8\uD558\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4
+MailValidatorDialog.ResultStringFalseField.Tooltip=\uC774\uBA54\uC77C\uC774 \uC720\uD6A8\uD558\uC9C0 \uC54A\uC73C\uBA74 \uD45C\uC2DC
