+++ plugins/mail-validator/impl/src/main/resources/org/pentaho/di/trans/steps/mailvalidator/messages/messages_zh_CN.properties
@@ -1,0 +1,93 @@
+#File generated by Hitachi Vantara Translator for package 'org.pentaho.di.trans.steps.MailValidator' in locale 'en_US'
+#
+#
+#Thu Jun 19 19:10:46 CEST 2008
+
+MailValidatorDialog.Shell.Title=\u68c0\u9a8c\u90ae\u4ef6\u5730\u5740
+MailValidatorDialog.FailedToGetFields.DialogMessage=Unable to get fields from previous steps because of an error
+MailValidatorMeta.Exception.UnexpectedErrorReadingStepInfo=Unexpected error reading step information from the repository
+MailValidatorDialog.Stepname.Label=\u6b65\u9aa4\u540d\u79f0
+MailValidator.Log.UnexpectedError=Unexpected error in ''
+MailValidatorDialog.emailFieldName.Label=\u90ae\u4ef6\u5b57\u6bb5\u540d 
+MailValidatorDialog.FailedToGetFields.DialogTitle=\u83b7\u53d6\u5b57\u6bb5\u5931\u8d25
+MailValidator.LineNumber=linenr {0}
+MailValidatorMeta.CheckResult.TableFieldOK=Table fieldname is provided
+MailValidator.ErrorInStepRunning=Because of an error, this step can''t continue\: 
+MailValidator.Log.ErrorFindingField=Error finding field\: 
+MailValidatorMeta.Exception.UnableToSaveStepInfo=Unable to save step information to the repository for id_step\=
+MailValidatorMeta.CheckResult.CouldNotReadFields=Couldn''t read fields from the previous step.
+
+MailValidatorMeta.CheckResult.NoInpuReceived=No input received from other steps\!
+MailValidator.Exception.CouldnotFindField=Couldn''t find field ''{0}'' in row\!
+MailValidatorMeta.CheckResult.ReceivingInfoFromOtherSteps=Step is receiving info from other steps.
+MailValidatorDialog.ResultField.Label=\u7ed3\u679c\u5b57\u6bb5\u540d
+MailValidatorDialog.ResultField.Tooltip=\u7ed3\u679c\u5b57\u6bb5\u540d
+
+MailValidator.Error.FilenameFieldMissing=Filename is missing\!
+MailValidatorMeta.CheckResult.ResultFieldMissing=Result fieldname is missing\!
+MailValidatorMeta.CheckResult.ResultFieldOk=\u7ed3\u679c\u5b57\u6bb5\u540d\u5df2\u6307\u5b9a
+
+MailValidatorMeta.CheckResult.ErrorOccurred=An error occurred\: 
+
+MailValidatorMeta.Exception.UnableToReadStepInfo=Unable to read step information from XML
+MailValidatorMeta.CheckResult.eMailFieldMissing=Email field name is missing\!
+MailValidatorMeta.CheckResult.eMailFieldOK=Email field name was specified
+MailValidatorDialog.ResultGroup.label=\u7ed3\u679c
+MailValidatorDialog.ResultAsString.Label=\u7ed3\u679c\u662f\u4e00\u4e2a\u5b57\u7b26\u4e32
+MailValidatorDialog.ResultAsString.Tooltip=\u7ed3\u679c\u662f\u4e00\u4e2a\u5b57\u7b26\u4e32
+MailValidatorDialog.ResultStringTrueField.Label=\u90ae\u4ef6\u5730\u5740\u6709\u6548
+MailValidatorDialog.ResultStringTrueField.Tooltip=display when email address is valid
+MailValidatorDialog.ResultStringFalseField.Label=\u90ae\u4ef6\u5730\u5740\u65e0\u6548
+MailValidatorDialog.ResultStringFalseField.Tooltip=display when email is not valid
+MailValidatorMeta.CheckResult.eMailValidMsgMissing=eMail valid message is missing!
+MailValidatorMeta.CheckResult.eMailValidMsgOk=eMail valid message was specified.
+MailValidatorMeta.CheckResult.eMailNotValidMsgMissing=eMail not valid message is missing!
+MailValidatorMeta.CheckResult.eMailNotValidMsgOk=eMail not valid message was specified.
+MailValidatorDialog.ErrorMsgField.Label=\u9519\u8bef\u5b57\u6bb5
+MailValidatorDialog.ErrorMsgField.Tooltip=\u9519\u8bef\u5b57\u6bb5
+MailValidator.Error.ResultFieldNameMissing=Result fieldname is missing or empty!
+MailValidatorDialog.SettingsGroup.Label=\u8bbe\u7f6e
+MailValidatorDialog.SMTPCheck.Label=SMTP \u68c0\u67e5
+MailValidatorDialog.SMTPCheck.Tooltip=\u9009\u4e2d\u6b64\u9009\u9879\uff0cPDI\u4f1a\u68c0\u67e5SMTP\uff0c\u5426\u5219\u9ed8\u8ba4\u503c\u68c0\u67e5\u90ae\u4ef6\u5730\u5740\u683c\u5f0f.
+MailValidatorDialog.eMailSenderField.Label=\u53d1\u4ef6\u4eba
+MailValidatorDialog.eMailSenderField.Tooltip=This address will be used when performing SMTP check.
+MailValidatorDialog.TimeOutField.Label=\u8d85\u65f6(\u6beb\u79d2)
+MailValidatorDialog.TimeOutField.Tooltip=In order to perform a SMTP check, PDI will open a socket.\nSpecify here the timeout in milliseconds.
+MailValidatorDialog.DefaultSMTPField.Label=\u9ed8\u8ba4 SMTP \u670d\u52a1\u5668
+MailValidatorDialog.DefaultSMTPField.Tooltip=When PDI perform a SMTP check,\nit extracts exhangers list from email's domain name.\nDepending on the size of the list, the process can last.\nIf you know which SMTP server to query, please specify it as default SMTP server\nin that case, PDI will query only this server.
+MailValidatorMeta.CheckResult.eMailSenderMissing=Email sender is missing!
+MailValidatorMeta.CheckResult.eMailSenderOk=Email sender was specified.
+MailValidatorDialog.dynamicDefaultSMTP.Label=\u52a8\u6001\u9ed8\u8ba4 SMTP?
+MailValidatorDialog.dynamicDefaultSMTP.Tooltip=Select this option if you want to get dynamically\ndefault SMTP server.You must select below the field that contain value.
+MailValidatorDialog.defaultSMTPField.Label=\u9ed8\u8ba4 SMTP \u5b57\u6bb5
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldMissing=Dynamic default SMTP field is missing!
+MailValidatorMeta.CheckResult.dynamicDefaultSMTPFieldOk=Dynamic default SMTP field was specified.
+MailValidator.Error.EmailValidMsgMissing=Email valid message is empty!
+MailValidatorMeta.CheckResult.EmailValidMsgOk=Email valid message was specified.
+MailValidator.Error.EmailNotValidMsgMissing=Email not valid message is empty!
+MailValidatorMeta.CheckResult.EmailNotValidMsgOk=Email not valid message was specified.
+MailValidator.Error.EMailValidMsgMissing=Email valid message is empty!
+MailValidator.Error.EMailNotValidMsgMissing=Email not valid message is empty!
+MailValidator.NoMatchName=No match for name '{0}' 
+MailValidator.MalformedAddress=Malformed address {0}!
+MailValidator.ClassName=\u90ae\u4ef6\u5730\u5740\u6821\u9a8c
+MailValidator.NoMachinesInDomain=We do not find machines for domain : {0}!
+MailValidator.ErrorGettingMachinesInDomain=Error getting machines : {0}
+MailValidator.ExchangersFound=We found {0} exchanger(s).
+MailValidator.TryingExchanger=Trying Exchanger : {0} ...
+MailValidator.InvalidHeader=Invalid header
+MailValidator.SayHello= 1\ufffd Let''s say Hello to server : EHLO {0}
+MailValidator.CheckSender= 2\ufffd Let''s check sender. MAIL FROM: <{0}> 
+MailValidator.ServerReplied=Server successfully replied : {0}
+MailValidator.SenderRejected=Sender rejected
+MailValidator.SenderAccepted=Sender accepted : {0}
+MailValidator.CheckReceiver=3\ufffd Let''s check receiver. RCPT TO: <{0}>
+MailValidator.AddressNotValid=Address  : {0} is not valid!
+MailValidator.ReceiverAccepted=Receiver {0} accepted : {1}
+
+MailValidator.Error.DefaultSMTPFieldMissing=Default SMTP server is missing!
+MailValidator.Error.ResultFieldMissing=Result fieldname is missing!
+MailValidatorMeta.CheckResult.EmailNotValidMsgMissing=EMail not valid message is missing!
+MailValidatorMeta.CheckResult.EmailValidMsgMissing=EMail valid message is missing!
+MailValidator.Error.MailEmpty=Email address is empty!
+MailValidator.ConnectingTo=Connecting to {0} on port {1} ...{2}
